requirements.txt
python-telegram-bot==20.4
python-dotenv==1.0.0
psutil==5.9.5
psycopg[binary]==3.2.13
requests==2.31.0
openpyxl==3.1.2
pandas==2.3.3
xlsxwriter==3.1.2
matplotlib==3.10.8
pillow==12.1.0
aiohttp==3.13.3
numpy==2.3.3






